requirements.txt
aws-cdk-lib==2.2.0
constructs>=10.0.0,<11.0.0
python-dotenv==0.19.2
cdk-ec2-key-pair==3.0.1
